Use void pointers, not (unsigned char *).
At the very least, u8 * is a better choice than working in terms of char, which has not always been historically significant to C as an 8-bit type.  If not for readability, even the flexibility of void pointers is more readable than assuming units of char that correspond between both MIPS and whatever host machine.