warning C4701: potentially uninitialized local variable 'block' used
Evidently, using a do..while loop instead of while or for loops fixes the false assumption that block can be potentially uninitialized.  Another fix is to set block = 0; just outside or before the loop, but that seems stupid since setting block to 0 is never part of the equation as a whole and only risks generating extra code on not-as-optimizing compilers unnecessarily.