just in case sizeof(i8) > sizeof(char)
sizeof(char) is always 1, obviously, but for primitive (and possibly even pre-ANSI) implementations of C, it may be possible that `char` is not an 8-bit type.  In the worst case scenario (impossible, though, with C89), char can be less than 8 bits, and we needed to typedef `u8` to something else like `unsigned short` to u8 and do bit masks all the time to fully support emulation of the MIPS memory model.

As extreme as that may sound, I see no repercussions in return for this portability.