In case of block read errors, only break the loop.
Returning out of the whole function exits ungracefully before we have a chance to close the file stream.  In case of block read errors, we still want to jump to the end of the function to close the file stream as we do in the event that no errors were encountered.